Add CancellationExtensions for async task handling with cancellation support
